remove nvml update thermal vision effect
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1527 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/client/cl_download.c
@@ -536,12 +536,12 @@ void CL_DownloadFileName(char* dest, int destlen, char* fn)
 {
 	if (strncmp(fn, "players", 7) == 0)
 	{
-		Com_sprintf(dest, destlen, "%s/%s", BASEDIRNAME, fn);
+		Com_sprintf(dest, (size_t)destlen, "%s/%s", BASEDIRNAME, fn);
 	}
 
 	else
 	{
-		Com_sprintf(dest, destlen, "%s/%s", FS_Gamedir(), fn);
+		Com_sprintf(dest, (size_t)destlen, "%s/%s", FS_Gamedir(), fn);
 	}
 }
 
@@ -766,7 +766,7 @@ CL_ParseDownload(void)
 		}
 	}
 
-	fwrite(net_message.data + net_message.readcount, 1, size, cls.download);
+	fwrite(net_message.data + net_message.readcount, 1, (size_t)size, cls.download);
 	net_message.readcount += size;
 
 	if (percent != 100)
